+++ test/integration/__fixtures__/signTx.ts
@@ -678,7 +678,7 @@ export const testsShelleyNoCertificates: TestcaseShelley[] = [
             network: Networks.Testnet,
             outputs: [outputs.externalShelleyScripthash],
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/0/0")],
         result: {
             txHashHex:
@@ -857,7 +857,7 @@ export const testsShelleyNoCertificates: TestcaseShelley[] = [
                 },
             ],
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/2/0")],
         txBody: "a500818258203b40265111d8bb3c3c608d95b3a0bf83461ace32d79336579a1939b3aa" +
       "d1c0b700018182582b82d818582183581c9e1c71de652ec8b85fec296f0685ca3988781c94a2e1" +
@@ -925,7 +925,7 @@ export const testsShelleyWithCertificates: TestcaseShelley[] = [
                 },
             ],
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/2/0")],
         txBody: "a500818258201af8fa0b754ff99253d983894e63a2b09cbb56c833ba18c3384210163" +
       "f63dcfc00018182582b82d818582183581c9e1c71de652ec8b85fec296f0685ca3988781c94a2" +
@@ -1001,7 +1001,7 @@ export const testsShelleyWithCertificates: TestcaseShelley[] = [
                 },
             ],
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/2/0")],
         txBody: "a500818258201af8fa0b754ff99253d983894e63a2b09cbb56c833ba18c3384210163" +
       "f63dcfc00018182582b82d818582183581c9e1c71de652ec8b85fec296f0685ca3988781c94a2" +
@@ -1079,7 +1079,7 @@ export const testsShelleyWithCertificates: TestcaseShelley[] = [
                 },
             ],
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/2/0")],
         txBody: "a500818258201af8fa0b754ff99253d983894e63a2b09cbb56c833ba18c3384210163" +
       "f63dcfc00018182582b82d818582183581c9e1c71de652ec8b85fec296f0685ca3988781c94a2" +
@@ -1311,7 +1311,7 @@ export const testsShelleyWithCertificates: TestcaseShelley[] = [
                 },
             },
         },
-        signingMode: TransactionSigningMode.SCRIPT_TRANSACTION,
+        signingMode: TransactionSigningMode.MULTISIG_TRANSACTION,
         additionalWitnessPaths: [str_to_path("1854'/1815'/0'/0/0"), str_to_path("1854'/1815'/0'/2/0")],
 
         result: {